# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: Palm API
  description: |
    RESTful API for Palm Application - Modular Monolith Architecture
    Implements CQRS pattern with clean separation between Identity and Sales modules.
  version: 1.0.0
  contact:
    name: Palm Development Team
    email: dev@palm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.palm.com/v1
    description: Production server
...
paths:
...
  /users/{id}/permissions:
    get:
      tags:
        - Users
      summary: Get user permissions
      description: Get the permissions assigned to a specific user
      operationId: getUserPermissions
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user_id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: John Doe
                      email:
                        type: string
                        example: john.doe@example.com
                      role:
                        type: string
                        example: admin
                      user_type:
                        type: string
                        example: employee
                      permissions:
                        type: array
                        items:
                          type: string
                        example: ['users.manage', 'roles.manage', 'rooms.manage']
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
        message:
          type: string
          example: The requested resource could not be found
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
security:
  - bearerAuth: []